split cli options, inference, option definitions, utils to own files
